Initial commit with files
Note: The liblua dynamic library for version 5.1 on MacOS (which the trial requires) is not built when installed via homebrew or when built from binaries provided by the official Lua website. This results in a linking error whenever trying to build the client. Although the liblua5.1.dll can be found online the .dylib cannot (at least from what I could tell from several days of searching). I'm building on a M1 macbook. If another person was able to build off the same architecture from scratch, then disregard, otherwise this could be an issue further down the line as more people who choose to work with MacOS try to build the otclient.